requirements.txt
fastapi==0.116.1
uvicorn[standard]==0.37.0
psycopg2-binary==2.9.10
sqlalchemy==2.0.43
alembic==1.16.5
python-dotenv==1.1.1
pydantic[email]==2.11.9
passlib[bcrypt]==1.7.4
bcrypt==4.0.1
python-jose[cryptography]==3.3.0